Update README.md
This story, an alternate take on the Taishō era where magic and spirits are real but fading, centers on Miyo Saimori.

Born without any supernatural abilities, Miyo is subjected to a life of abusive servitude by her stepmother. When she is finally old enough to be married off, her dreams of escaping her miserable life are shattered. Her fiancé is Kiyoka Kudou, a handsome military commander rumored to be so cold and cruel that his previous brides all fled within three days.

Having nowhere else to go, Miyo prepares for a terrible life, but soon discovers that her prospective husband is not the monster everyone said he was. As Miyo and Kiyoka slowly begin to trust and connect with each other, they realize that their unexpected relationship might just be the path to true love and happiness for both of them.